refactor: shift instance admin restriction content to seperate component. fix: instance components export logic.
