+++ web/components/instance/index.ts
@@ -2,3 +2,9 @@ export * from "./help-section";
 export * from "./sidebar-menu";
 export * from "./sidebar-dropdown";
 export * from "./general-form";
+export * from "./ai-form";
+export * from "./email-form";
+export * from "./github-config-form";
+export * from "./google-config-form";
+export * from "./image-config-form";
+export * from "./instance-admin-restriction";
